API documentation

Types and constructors

BSplines.AllDerivatives — Type

AllDerivatives{N}

A singleton type that represents all m-th derivatives where 0 ≤ m < N.

BSplines.AllDerivatives — Method

AllDerivatives(N)

A shortcut for AllDerivatives{N}(), representing all m-th derivatives where 0 ≤ m < N.

BSplines.BSpline — Type

BSpline{B} <: Spline{B}

Type for a B-spline from a B-spline basis of type B. BSplines can be obtained by indexing into a B-spline basis.

BSplines.BSpline — Method

BSpline(basis::BSplineBasis, index)

Return the index-th B-spline of basis.

Example

julia> basis = BSplineBasis(5, 1:10);

julia> BSpline(basis, 3)
BSpline{BSplineBasis{UnitRange{Int64}}}:
 basis: 13-element BSplineBasis{UnitRange{Int64}}:
  order: 5
  breakpoints: 1:10
 index: 3 (knots: [1, 1, 1, 2, 3, 4])

BSplines.BSplineBasis — Type

BSplineBasis{T<:AbstractVector{<:Real}}

Type for a B-spline basis with breakpoint vector of type T.

Here, a B-spline basis is completely specified by its order $k$ and breakpoint sequence. The knot sequence is derived from the breakpoint sequence by duplicating the first and last breakpoints so they each appear $k$ times. Knot sequences where the first and last breakpoints do not appear $k$ times are not supported by this data type.

BSplines.BSplineBasis — Method

BSplineBasis(order, breakpoints)

Create a B-spline basis with order order and breakpoint vector breakpoints. The breakpoint vector is assumed to be sorted.

BSplines.Derivative — Type

Derivative{N}

A singleton type that represents the N-th derivative.

BSplines.Derivative — Method

Derivative(N)

A shortcut for Derivative{N}(), representing the N-th derivative.

BSplines.Spline — Type

Spline{B<:BSplineBasis, C<:AbstractVector{<:Real}}

Type for a spline based on a B-spline basis of type B and coefficient vector of type C.

BSplines.Spline — Method

Spline(basis, coeffs)

Create a spline from a B-spline basis and a vector of coefficients.

Core.Function — Type

Function(spline::Spline, [deriv::Derivative], [onlysupport=true])

Wrap spline in an object that is a subtype of Function. Calling the returned function with a single argument x will evaluate the spline (or one of its derivatives as specified by deriv) at x.

If the optional argument onlysupport is set to true (the default), the returned function will return NaN if evaluated outside of the support of spline. If onlysupport is set to false, it will return zero there (as does calling the spline directly).

Note that a Spline can be called without wrapping them as described here, although they are not a subtype of Function. Wrapping a spline in a Function object is mainly intended to aid in plotting, which is the rationale behind the onlysupport=true default: when using Plots.jl, this will cause the spline to not be drawn outside of its support.

Examples

julia> spline = approximate(sin, BSplineBasis(5, 0:5)); # create a Spline

julia> f = Function(spline, false); # f is zero outside of the interval [0,5]

julia> g = Function(spline, Derivative(1)); # g is NaN outside of the interval [0,5]

julia> f(1.5) == spline(1.5)
true

julia> g(1.5) == spline(1.5, Derivative(1))
true

julia> f(-1), g(-1)
(0.0, NaN)

Functions

BSplines.approximate — Function

approximate(f, basis::BSplineBasis; indices=eachindex(basis)) -> Spline

Approximate the function f in the B-spline basis basis. If indices is supplied, only the basis functions at the given indices of basis are used.

The approximation is calculated by interpolating samples of f at the knotaverages of the basis.

See also: interpolate

Examples

julia> basis = BSplineBasis(5, 0:5);

julia> spl = approximate(sin, basis, indices=2:length(basis))
Spline{BSplineBasis{UnitRange{Int64}},Array{Float64,1}}:
 basis: 9-element BSplineBasis{UnitRange{Int64}}:
  order: 5
  breakpoints: 0:5
 coeffs: [0.0, 0.24963094468700395, 0.7525104872191076, 1.229735980709192, 0.7385208497317045, -0.4328168377896504, -1.0125409246826416, -1.029692234224304, -0.9589242746631385]

julia> spl(π/4)
0.7071028397621081

BSplines.averagebasis — Function

averagebasis(order, datapoints) -> BSplineBasis

Returns a B-spline basis with the specified order that is well-suited for interpolation on the given datapoints. The datapoints vector is assumed to be sorted.

The calculated breakpoints are described in [deBoor1978], p. 219, as a “reasonable alternative” to the optimal breakpoint sequence since they are “often very close to the optimum” and are computationally inexpensive.

Examples

julia> averagebasis(5, 0:10)
11-element BSplineBasis{Array{Float64,1}}:
 order: 5
 breakpoints: [0.0, 2.5, 3.5, 4.5, 5.5, 6.5, 7.5, 10.0]

BSplines.basis — Function

basis(spline::Spline)

Return the B-spline basis on which the spline is defined.

BSplines.breakpoints — Function

breakpoints(basis::BSplineBasis)

Return the breakpoint sequence of the B-spline basis.

Examples

julia> breakpoints(BSplineBasis(3, 0:5))
0:5

julia> breakpoints(BSplineBasis(4, [1.0, 1.5, 2.5, 4.0]))
4-element Array{Float64,1}:
 1.0
 1.5
 2.5
 4.0

BSplines.bsplines! — Function

bsplines!(dest, basis, x; leftknot=intervalindex(basis, x)) -> offset

Calculate the values of all B-splines of basis that are non-zero at x and store the result in dest. The destination vector dest must have the length order(basis).

If any B-splines are non-zero at x, an integer offset is returned and the value of the i-th B-spline is written to dest[i-offset]. If no B-splines are non-zero at x, nothing is returned and dest is not mutated.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> dest = zeros(4);

julia> bsplines!(dest, BSplineBasis(4, 0:5), 2.4)
2

julia> dest # dest[i] contains value of (i+2)-th B-spline
4-element Array{Float64,1}:
 0.03600000000000002
 0.5386666666666667 
 0.41466666666666663
 0.01066666666666666

bsplines!(dest, basis, x, ::Derivative{N}; leftknot=intervalindex(basis, x))

Calculate the values of all B-splines of basis that are non-zero at x and store the result in dest. The destination vector dest must have the length order(basis).

If any B-splines are non-zero at x, an integer offset is returned and the N-th derivative of the i-th B-spline is written to dest[i-offset]. If no B-splines are non-zero at x, nothing is returned and dest is not mutated.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> dest = zeros(4);

julia> bsplines!(dest, BSplineBasis(4, 0:5), 7.0, Derivative(2)) # returns nothing

julia> bsplines!(dest, BSplineBasis(4, 0:5), 4.2, Derivative(2))
4

julia> dest # dest[i] contains 2nd derivative of (i+4)-th B-spline
4-element Array{Float64,1}:
  0.7999999999999998
 -1.399999999999999 
 -0.6000000000000019
  1.200000000000001 

bsplines!(dest, basis, x, ::AllDerivatives{N}; leftknot=intervalindex(basis, x))

Calculate the values of all B-splines of basis that are non-zero at x and store the result in dest. The destination matrix dest must have the dimensions order(basis)×N.

If any B-splines are non-zero at x, an integer offset is returned and the m-th derivative of the i-th B-spline is written to dest[i-offset, m+1]. If no B-splines are non-zero at x, nothing is returned and dest is not mutated.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> dest = zeros(4, 3);

julia> bsplines!(dest, BSplineBasis(4, 0:5), -1.0, AllDerivatives(3)) # returns nothing

julia> bsplines!(dest, BSplineBasis(4, 0:5), 3.75, AllDerivatives(3))
3

julia> dest # dest[i,m] contains (m-1)-th derivative of (i+3)-th B-spline
4×3 Array{Float64,2}:
 0.00260417  -0.03125    0.25 
 0.315104    -0.65625    0.25 
 0.576823     0.265625  -1.625
 0.105469     0.421875   1.125

BSplines.bsplines — Function

bsplines(basis, x; leftknot=intervalindex(basis, x))

Calculate the values of all non-zero B-splines of basis at x.

If any B-splines are non-zero at x, an OffsetVector is returned that contains the value of the i-th B-spline at the index i. If no B-splines are non-zero at x, nothing is returned.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> bsplines(BSplineBasis(4, 0:5), 2.4)
4-element OffsetArray(::Array{Float64,1}, 3:6) with eltype Float64 with indices 3:6:
 0.03600000000000002
 0.5386666666666667 
 0.41466666666666663
 0.01066666666666666

julia> bsplines(BSplineBasis(4, 0:5), 6) # returns nothing

julia> bsplines(BSplineBasis(3, 0:5), 17//5, leftknot=6)
3-element OffsetArray(::Array{Rational{Int64},1}, 4:6) with eltype Rational{Int64} with indices 4:6:
  9//50
 37//50
  2//25

bsplines(basis, x, ::Derivative{N}; leftknot=intervalindex(basis, x))

Calculate the N-th derivatives of all B-splines of basis that are non-zero at x.

If any B-splines are non-zero at x, an OffsetVector is returned that contains the N-th derivative of the i-th B-spline at the index i. If no B-splines are non-zero at x, nothing is returned.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> bsplines(BSplineBasis(3, 0:5), 2.4, Derivative(1))
3-element OffsetArray(::Array{Float64,1}, 3:5) with eltype Float64 with indices 3:5:
 -0.6000000000000001 
  0.20000000000000018
  0.3999999999999999 

julia> bsplines(BSplineBasis(3, 0:5), 6, Derivative(1)) # returns nothing

julia> bsplines(BSplineBasis(4, 0:5), 17//5, Derivative(2), leftknot=7)
4-element OffsetArray(::Array{Rational{Int64},1}, 4:7) with eltype Rational{Int64} with indices 4:7:
  3//5
 -4//5
 -2//5
  3//5

bsplines(basis, x, ::AllDerivatives{N}; leftknot=intervalindex(basis, x))

Calculate all m-th derivatives (0 ≤ m < N) of all B-splines of basis that are non-zero at x.

If any B-splines are non-zero at x, an OffsetMatrix is returned that contains the m-th derivative of the i-th B-spline at the index i, m. If no B-splines are non-zero at x, nothing is returned.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Examples

julia> bsplines(BSplineBasis(3, 0:5), 2.4, AllDerivatives(3))
3×3 OffsetArray(::Array{Float64,2}, 3:5, 0:2) with eltype Float64 with indices 3:5×0:2:
 0.18  -0.6   1.0
 0.74   0.2  -2.0
 0.08   0.4   1.0

julia> bsplines(BSplineBasis(3, 0:5), 6.0, AllDerivatives(3)) # returns nothing

julia> bsplines(BSplineBasis(4, 0:5), 17//5, AllDerivatives(4), leftknot=7)
4×4 OffsetArray(::Array{Rational{Int64},2}, 4:7, 0:3) with eltype Rational{Int64} with indices 4:7×0:3:
   9//250   -9//50   3//5  -1//1
 202//375  -14//25  -4//5   3//1
 307//750   31//50  -2//5  -7//2
   2//125    3//25   3//5   3//2

BSplines.coeffs — Function

coeffs(spline::Spline)

Return the coefficient vector of spline.

BSplines.interpolate — Function

interpolate(basis::BSplineBasis, xvalues, yvalues; indices=eachindex(basis)) -> Spline

Interpolate the data given by xvalues and yvalues in the B-spline basis basis. If indices is supplied, only the basis functions at the given indices of basis are used.

The spline interpolation is calculated by creating the matrix B = [basis[i](x) for x=xvalues, i=indices] and then calculating B\yvalues.

See also: approximate

Examples

julia> basis = BSplineBasis(5, 1:10);

julia> xs = range(1, stop=10, length=length(basis)); ys = log.(xs);

julia> spl = interpolate(basis, xs, ys)
Spline{BSplineBasis{UnitRange{Int64}},Array{Float64,1}}:
 basis: 13-element BSplineBasis{UnitRange{Int64}}:
  order: 5
  breakpoints: 1:10
 coeffs: [0.0, 0.2480193113006778, 0.5968722382485209, 0.9466707785821219, 1.2689430722820556, 1.5140484163988175, 1.7114875128056504, 1.8766572742486973, 2.0185639127626325, 2.1429230073972407, 2.2259183994808813, 2.2775846911059, 2.302585092994046]

julia> spl(float(ℯ))
0.9999766059171411

BSplines.intervalindex — Function

intervalindex(vec, x[, start])

If v is an AbstractVector, return the largest index i so that vec[i] ≤ x and vec[i] < vec[end]. Return nothing if x < first(vec) or x > last(vec) or isnan(x). The vector vec is assumed to be sorted in ascending order.

If vec is a BSplineBasis, return intervalindex(knots(vec), x[, start]).

If start is given, a linear search is performed, starting from the index start going forward or backward. If start is not given, a binary search is performed.

Examples

julia> intervalindex([1,1,2,3,4,4,4,5,6,6], 2.5)
3

julia> intervalindex([1,1,2,3,4,4,4,5,6,6], 7) # returns nothing

julia> intervalindex([1,1,2,3,4,4,4,5,6,6], 1)
2

julia> intervalindex([1,1,2,3,4,4,4,5,6,6], 4)
7

julia> intervalindex([1,1,2,3,4,4,4,5,6,6], 6.0)
8

BSplines.intervalindices — Function

intervalindices(basis::BSplineBasis, indices=eachindex(basis))

Return an iterator that yields the indices of all intervals on which basis is defined, i.e., it produces all indices ind (in ascending order) for which (knots(basis)[ind], knots(basis)[ind+1]) is such an interval.

If a range of indices is supplied, the iterator yields only those intervals on which at least one of the B-splines basis[j] for j=indices is non-zero.

Examples

julia> intervalindices(BSplineBasis(3, 0:5))
3:7

julia> intervalindices(BSplineBasis(3, 0:5), 1:4)
3:6

julia> intervalindices(BSplineBasis(4, [1,2,3,4,4,4,5,6]))
BSplines.IntervalIndices{Array{Int64,1}}([1, 2, 3, 4, 4, 4, 5, 6], 1:8, 3)

julia> collect(ans)
5-element Array{Int64,1}:
  4
  5
  6
  9
 10

intervalindices(basis::BSplineBasis, i, j, ...)

For integers i, j, …, return an iterator that yields the indices of all intervals on which all of the B-splines basis[i], basis[j], … are non-zero, i.e., it produces all indices ind (in ascending order) for which (knots(basis)[ind], knots(basis)[ind+1]) is such an interval.

Examples

julia> intervalindices(BSplineBasis(3, 0:5), 3)
3:5

julia> intervalindices(BSplineBasis(3, 0:5), 4, 5)
5:6

julia> intervalindices(BSplineBasis(3, 0:5), 2, 6) # B-splines do not overlap
6:5

julia> intervalindices(BSplineBasis(3, 0:5), 3, 5, 4)
5:5

julia> intervalindices(BSplineBasis(4, [1,2,3,4,4,4,5,6]), 3, 5)
BSplines.IntervalIndices{Array{Int64,1}}([1, 2, 3, 4, 4, 4, 5, 6], 2:4, 3)

julia> collect(ans)
2-element Array{Int64,1}:
 5
 6

BSplines.knotaverages! — Function

knotaverages!(dest, basis::BSplineBasis; indices=eachindex(basis))

Calculate the knot averages τ[i] = mean(knots[i+1:i+order-1]) for i ∈ indices and the knots and order of basis and store the result in dest. The knot averages are recommended in [deBoor1978] (p. 214) as data points for interpolation.

See also: knotaverages

Examples

julia> dest = Vector{Float64}(undef, 7);

julia> knotaverages!(dest, BSplineBasis(3, 0:5))
7-element Array{Float64,1}:
 0.0
 0.5
 1.5
 2.5
 3.5
 4.5
 5.0

julia> dest = Vector{Rational{Int}}(undef, 5);

julia> knotaverages!(dest, BSplineBasis(3, 0:5), indices=2:6)
5-element Array{Rational{Int64},1}:
 1//2
 3//2
 5//2
 7//2
 9//2

BSplines.knotaverages — Function

knotaverages(basis::BSplineBasis; indices=eachindex(basis))

Return the knot averages τ[i] = mean(knots[i+1:i+order-1]) for i ∈ indices and the knots and order of basis. The knot averages are recommended in [deBoor1978] (p. 214) as data points for interpolation.

See also: knotaverages!

Examples

julia> knotaverages(BSplineBasis(3, 0:5))
7-element Array{Float64,1}:
 0.0
 0.5
 1.5
 2.5
 3.5
 4.5
 5.0

julia> knotaverages(BSplineBasis(4, [1, 3//2, 5//2, 4]), indices=2:6)
5-element Array{Rational{Int64},1}:
 7//6
 5//3
 8//3
 7//2
 4//1

BSplines.knots — Function

knots(basis::BSplineBasis)

Return the knot sequence of the B-spline basis.

The knot sequence is the breakpoint sequence except that the first and last values are duplicated so they appear order(basis) times.

Examples

julia> knots(BSplineBasis(3, 0:5))
10-element BSplines.KnotVector{Int64,UnitRange{Int64}}:
 0
 0
 0
 1
 2
 3
 4
 5
 5
 5

BSplines.order — Function

order(spline::Spline)
order(basis::BSplineBasis)

Return the order of a spline or a B-spline basis.

Examples

julia> order(BSplineBasis(3, 0:5))
3

julia> order(BSplineBasis(4, [1.0, 1.5, 2.5, 4.0]))
4

BSplines.splinevalue — Function

splinevalue(spline::Spline, x; leftknot=intervalindex(basis(spline), x))

Calculate the values of spline at x.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Instead of calling splinevalue, a spline object can be called directly: spline(x; [leftknot]) is equivalent to splinevalue(spline, x; [leftknot]).

Examples

julia> spl = Spline(BSplineBasis(4, 0:5), 1:8);

julia> splinevalue(spl, 1.7)
3.69775

julia> splinevalue(spl, 3.6, leftknot=7)
5.618

julia> spl(18//5)
2809//500

julia> spl(5, leftknot=8)
8.0

splinevalue(spline::Spline, x, ::Derivative{N}; leftknot=intervalindex(basis(spline), x))

Calculate the value of the N-th derivative of spline at x.

If the index of the relevant interval is already known, it can be supplied with the optional leftknot keyword to speed up the calculation.

Instead of calling splinevalue, a spline object can be called directly: spline(x, Derivative(N); [leftknot]) is equivalent to splinevalue(spline, x, Derivative(N); [leftknot]).

Examples

julia> spl = Spline(BSplineBasis(4, 0:5), 1:8);

julia> splinevalue(spl, 1.7, Derivative(1))
1.0225

julia> splinevalue(spl, 18//5, Derivative(2), leftknot=7)
3//10

julia> spl(3.6, Derivative(3))
0.5

julia> spl(5, Derivative(1))
3.0

BSplines.support — Function

support(basis::BSplineBasis) -> a, b

Return the interval $[a,b]$ on which the B-spline basis is defined, i.e., a is the first and b the last breakpoint of the basis.

Examples

julia> support(BSplineBasis(3, 0:5))
(0, 5)

julia> support(BSplineBasis(4, [1.0, 1.5, 2.5, 4.0]))
(1.0, 4.0)

support(spline::Spline) -> a, b

If spline is a BSpline, return the interval $[a,b]$ on which the B-spline is non-zero. Otherwise, return support(basis(spline)).

Examples

julia> basis = BSplineBasis(3, 0:5);

julia> spline = Spline(basis, ones(7));

julia> zerospline = Spline(basis, zeros(7));

julia> support(spline)
(0, 5)

julia> support(zerospline) # even though the spline is zero everywhere
(0, 5)

julia> support(basis[4]) # for BSplines, return their actual support
(1, 4)

deBoor1978Carl de Boor, A Practical Guide to Splines, New York, N.Y.: Springer, 1978.

deBoor1978Carl de Boor, A Practical Guide to Splines, New York, N.Y.: Springer, 1978.

deBoor1978Carl de Boor, A Practical Guide to Splines, New York, N.Y.: Springer, 1978.